Removed a warning, and added better interop with authentication service

+++ src/app/normad/recipes/detail/detail.component.ts
@@ -5,23 +5,26 @@ import { ActivatedRoute, Router } from '@angular/router';
 import { RecipeService } from '../../services/recipe.service';
 import { IIngredient } from '../../models/ingredient';
 import { IngredientService } from '../../services/ingredient.service';
+import { AuthenticationService } from '../../services/auth/authentication.service';
 
 @Component({
   selector: 'app-detail',
   templateUrl: './detail.component.html',
-  styleUrls: ['./detail.component.scss']
+  styleUrls: [ './detail.component.scss' ]
 })
 export class DetailComponent implements OnInit, OnDestroy {
 
   recipe: IRecipe;
   ingredients: IIngredient[];
   private sub: Subscription;
+  isAdmin: boolean;
 
   constructor(
     private _route: ActivatedRoute,
     private _router: Router,
     private _recipeService: RecipeService,
     private _ingredientService: IngredientService,
+    private _authenticationService: AuthenticationService
   ) { }
 
   ngOnInit() {
@@ -32,6 +35,8 @@ export class DetailComponent implements OnInit, OnDestroy {
       this._recipeService.get(id)
         .subscribe(ingredient => this.recipe = ingredient);
     });
+    this._authenticationService.isAdmin()
+      .subscribe(x => this.isAdmin = x);
   }
 
   getIngredient(id) {
@@ -43,9 +48,11 @@ export class DetailComponent implements OnInit, OnDestroy {
   }
 
   delete(): void {
-    this._recipeService.delete(this.recipe)
-      .subscribe(res => {
-        this._router.navigate(['/recipes']);
-      });
+    if (confirm(`Are you certain you want to delete the recipe: ${this.recipe.name}?`)) {
+      this._recipeService.delete(this.recipe)
+        .subscribe(res => {
+          this._router.navigate([ '/recipes' ]);
+        });
+    }
   }
 }
